backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.3
selenium>=4.15
webdriver-manager>=4.0
python-dotenv>=1.0
celery>=5.3
redis>=5.0
pymongo>=4.6
gunicorn>=21.0
gevent>=23.0
whitenoise>=6.6
